requirements.txt
azure-core==1.32.0
azure-identity==1.19.0
azure-keyvault-secrets==4.9.0
blinker==1.9.0
cachelib==0.9.0
certifi==2024.12.14
cffi==1.17.1
charset-normalizer==3.4.1
click==8.1.8
colorama==0.4.6
contourpy==1.3.1
cryptography==44.0.0
cycler==0.12.1
dash==2.18.2
dash-bootstrap-components==1.6.0
dash-core-components==2.0.0
dash-html-components==2.0.0
dash-table==5.0.0
findlibs==0.0.5
flask==3.0.3
flask-caching==2.3.0
fonttools==4.55.3
greenlet==3.1.1
idna==3.10
importlib-metadata==8.5.0
isodate==0.7.2
itsdangerous==2.2.0
jinja2==3.1.5
kiwisolver==1.4.8
markupsafe==3.0.2
matplotlib==3.10.0
msal==1.31.1
msal-extensions==1.2.0
nest-asyncio==1.6.0
networkx==3.4.2
numpy==2.2.1
packaging==24.2
pandas==2.2.3
pillow==11.1.0
plotly==5.24.1
portalocker==2.10.1
pycountry==24.6.1
pycparser==2.22
pyjwt==2.10.1
pyodbc==5.2.0
pyodc==1.4.1
pyparsing==3.2.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
pytz==2024.2
pywin32==308
requests==2.32.3
retrying==1.3.4
scipy==1.15.1
setuptools==75.8.0
six==1.17.0
sqlalchemy==2.0.37
tenacity==9.0.0
typing-extensions==4.12.2
tzdata==2024.2
urllib3==2.3.0
werkzeug==3.0.6
wordcloud==1.9.4
zipp==3.21.0
